// plugins/woocommerce/client/blocks/assets/js/editor-components/search-list-control/types.ts
export type SearchListItem< T extends object = object > = {
	breadcrumbs: string[];
	children?: SearchListItem< T >[];
	count?: number;
	details?: T;
	id: string | number;
	name: string;
	parent: number;
	value: string;
};